+++ Lilu/Headers/kern_config.hpp
@@ -9,7 +9,11 @@
 #define kern_config_hpp
 
 // As for now claim to support all the features
-#define KEXTPATCH_SUPPORT 1
-#define COMPRESSION_SUPPORT 1
+#define LILU_KEXTPATCH_SUPPORT 1
+#define LILU_COMPRESSION_SUPPORT 1
+
+// Use these in plugins in Xcode Project Preprocessor Macros section
+// #define LILU_CUSTOM_IOKIT_INIT 1
+// #define LILU_CUSTOM_KMOD_INIT 1
 
 #endif /* kern_config_hpp */