+++ Lilu/PrivateHeaders/kern_config.hpp
@@ -17,13 +17,14 @@ class Configuration {
 	/**
 	 *  Possible boot arguments
 	 */
-	static constexpr const char *bootargOff {"-liluoff"};		// Disable the kext
-	static constexpr const char *bootargBeta {"-lilubeta"};		// Force enable the kext on unsupported os
-	static constexpr const char *bootargForce {"-liluforce"};   // Force enable the kext (including safe mode)
-	static constexpr const char *bootargDebug {"-liludbg"};		// Enable debug logging
-	static constexpr const char *bootargSlow {"-liluslow"};		// Prefer less destructive userspace measures
-	static constexpr const char *bootargFast {"-lilufast"};		// Prefer faster userspace measures
-	static constexpr const char *bootargLowMem {"-lilulowmem"};	// Disable decompression
+	static constexpr const char *bootargOff {"-liluoff"};			// Disable the kext
+	static constexpr const char *bootargBeta {"-lilubeta"};			// Force enable the kext on unsupported os
+	static constexpr const char *bootargBetaAll {"-lilubetaall"};	// Force enable the kext and all plugins on unsupported os
+	static constexpr const char *bootargForce {"-liluforce"};		// Force enable the kext (including safe mode)
+	static constexpr const char *bootargDebug {"-liludbg"};			// Enable debug logging
+	static constexpr const char *bootargSlow {"-liluslow"};			// Prefer less destructive userspace measures
+	static constexpr const char *bootargFast {"-lilufast"};			// Prefer faster userspace measures
+	static constexpr const char *bootargLowMem {"-lilulowmem"};		// Disable decompression
 	
 	/**
 	 * Minimal required kernel version
@@ -101,6 +102,11 @@ public:
 	 *  Install or recovery
 	 */
 	bool installOrRecovery {false};
+	
+	/**
+	 *  Beta for all plugins and Lilu itself
+	 */
+	bool betaForAll {false};
 	
 	/**
 	 *  Initialisation status