+++ Changelog.md
@@ -5,6 +5,7 @@ Lilu Changelog
 - Merged advanced disassembly API (thx Pb and others)
 - Added HDE disassembler for quick instruction decoding (by Vyacheslav Patkov)
 - Updated capstone to 3.0.5 rc3
+- Fixed load API lock type preventing dynamic memory allocation (thx Pb)
 
 #### v1.1.6
 - Ignored disabled kexts earlier for speed reasons