+++ Changelog.md
@@ -9,6 +9,7 @@ Lilu Changelog
 - Added setInterrupts API
 - Added an option to define custom plugin entry points
 - Added const reference evector API
+- Added FAT_CIGAM Mach-O support
 
 #### v1.1.6
 - Ignored disabled kexts earlier for speed reasons