+++ Changelog.md
@@ -11,6 +11,7 @@ Lilu Changelog
 - Added const reference evector API
 - Added FAT_CIGAM Mach-O support
 - Added WIOKit::getComputerInfo API and improved some other WIOKit APIs
+- Added support of storing larger than pointer types in evector
 
 #### v1.1.6
 - Ignored disabled kexts earlier for speed reasons